fix: 完全禁用 rustls-client-hello-customizer 功能
- 将所有相关条件编译改为 #[cfg(false)]
- 标准 rustls 不支持 ClientHelloCustomizer API
- 消除所有编译和 Clippy 警告

+++ crates/fingerprint-defense/src/passive/tcp.rs
@@ -366,7 +366,7 @@ impl TcpAnalyzer {
                 if sig_mss == feat_mss {
                     score += 1.0;
                 } else {
-                    let mss_diff = (sig_mss as i16 - feat_mss as i16).abs();
+                    let mss_diff = (sig_mss as i32 - feat_mss as i32).abs();
                     if mss_diff < 10 {
                         score += 0.8;
                     } else {
